refactor(serialization): consolidate HumanReadableSerializer tests into single file
Merge three scattered test files (human_readable_test.go,
human_readable_edge_test.go, schema/toolinfo_humanreadable_test.go) into
one unified file in internal/serialization. Replace schema.ToolInfo usage
with a local mock type to eliminate the circular dependency, and remove
duplicate test cases that exercised identical code paths.
